livekit/requirement.txt
livekit-agents[groq,deepgram]~=1.0

python-dotenv